[MAIN-FRAGMENT] activity fragment로 변경
