+++ app/src/main/java/com/hegunhee/newsimplememoapp/ui/main/MainActivity.kt
@@ -4,7 +4,6 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.setupWithNavController
 import com.hegunhee.newsimplememoapp.R
 import com.hegunhee.newsimplememoapp.databinding.ActivityMainBinding
 import dagger.hilt.android.AndroidEntryPoint
