+++ app/src/main/res/layout/fragment_detail_memo.xml
